Simplify rules for placement and titlebars on/off
In the past, using `titlebars_enabled = false` would make AwesomeWM create a hidden
empty titlebar on the client. Now it correctly creates the titlebar and
hides it.

Also now we do not need to delay the client placement function to
enforce correct placement since titlebars are correctly initialized
before the client is moved.